// backend_nodeJs/ms-game/src/models/cardDTO.ts
/**
 * Generated by orval v6.10.3 🍺
 * Do not edit manually.
 * Card Rest Api
 * Information about the Card Rest API and how to interact with
 * OpenAPI spec version: 1.0
 */

import { CardBasicsDTO } from "../../../../client/src/entities/cardBasicsDTO";

export interface CardDTO {
  id: number;
  model: CardBasicsDTO;
  userId: number;
  energy: number;
  hp: number;
  defence: number;
  attack: number;
  price: number;
}